backend/pyproject.toml
[tool.poetry]
name = "webapp-template"
version = "1.0.0"
description = "A FastAPI app connected to MongoDB"
authors = ["Maximilian Neudert <max@neudert.me>"]
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = "^0.32.0"
pydantic = "^2.10"
python-dotenv = "^1.0.0"
motor = "^3.6"
pandas = "^2.2"
polars = "^1.15"
pydantic-settings = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
httpx = "^0.27"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
